requirements.txt
aiogram>=3.11.0
Pillow>=10.3.0
python-dotenv>=1.0.0
